requirements.txt
transformers
torch
Pillow
streamlit
jsonlines
